Provide a parsedValue method for ModelFact
When using Arelle to parse and extract data from an XBRL instance, most of the time what you need is a Python value that you can then manipulate (eg. decimal.Decimal for numbers or True/False for booleans) rather than the string value which is what effectiveValue provides. The parsedValue method adds support for exactly that.