Fix numeric duplicate fact check to fire for any facts with same context, unit, concept, and decimals
The original code wouldn't fire a duplicate fact error unless all facts with the same context, unit, and concept had the same values. This commit changes the check to group facts with the same context, unit, and concept into groups based on the decimals. If we find multiple facts with the same decimals then we preceed with checking the values. This method of checking allows arelle to correctly identify all duplicate facts and fire errors for them