Don't allow fact aspect cache to grow indefinitely
For the vast majority of documents the default max size is never reached, but for a few massively dimensional reports (facts with thousands of dimensions) the cache can grow exponentially.

Other cache eviction mechanisms were investigated (LFU, LRU, etc.), but none were as performant as simply capping the cache size and not inserting anymore items after the max is reached.