Correction to `endDate` and `instantDate` helper functions
Made some changes to `datetimeValue` to accomodate `subtractOneDay`.
Also slightly reorganized the code for better flow. `hour24` was only
used in the "has time portion" if-statement and `addOneDay` in the
"else", so I moved them accordingly.
Added `subtractOneDay` as the last argument instead of intuitively next
to `addOneDay` since I'm not confident all uses of the function would
have used named arguments.