Validate Standard Taxonomies
Arelle currently skips full DTS validation for documents which are not
considered part of an extension taxonomy. However, this can lead to
undetected errors when a report references a broken taxonomy.

This change enables validation for all DTS documents, but continues to
not call the plugin validate DTS hook on base taxonomy documents which
would cause extension taxonomy rules to be applied to base taxonomies.