Tidy up comment on innerText()'s strip = True option
When strip = True (the default, then innerText assembles all the text as a single string and removes leading and trailing [ \t\r\n]. It does not do anything else, such as collapse runs of space.

This is a different style of normalization to any of XML whiteSpaces's so make that clear in the comment. What it does share with xs:whiteSpace is the set of characters it operates on.