Make use of xValue when viewing non-numeric facts
xValue contains the PSVI fact value. That means any xs:whitespace normlization will have been applied.

Make use of xValue in ModelFact.effectiveValue(), falling back to using value as is the usual pattern.

Make use of xValue in ModelInlineValueObject for non-numeric facts, again failling back to rawValue as is the usual pattern.

Change the calls to effectiveValue.strip()  to just effectiveValue in ModelFact.propertyView, ViewFileFactList.viewFacts and ViewWinFactList.viewFacts so that non-numeric facts are displayed correctly. That is, if they are supposed to have leading and trailing whitespace, then make sure they do have leading and trailing whitespace.