Reset disclosure system config on disabled plugin
If the user disables a plugin that had a selected disclosure system
the user continues to get an invalid disclosure system name error until
they enable a plugin with disclosure systems and select one. This change
resets the disclosure system name in the saved config if it's no longer
valid.