Change to a much more performant conversion from an LEI string to an int to checksum.
Rather than splitting the lei into chars, using a dictionary to lookup each character in the LEI string, then joining it back together ... define a translation table up front _leiToDigitTable and make use of str.translate() to do all the translation in one method call.

Also, rather having a hand-written conversion dict for each character, generate it programatically from string.ascii_uppercase, remembering we don't need to translate 0-9 to themselves, we just need to translate A-Z to 10 - 35.

We get an order of magnitude speed up from this so reduce the test time bounds to be between 50 milliseconds and 1 second.