Stop publishing tests as top level import
Installing the Arelle Python package makes both arelle.* and tests.*
packages available. This is a mistake. Consumers of the package don't
want our tests and it could easily conflict with their own packages.
This also reduces the size of the built wheel as tests are no longer
included. They are however, still included in the standard distribution
as expected.